#2302 - Introduce a low-level CAS storage interface
- Introduce CasStorageDriver interface
- Extract relevant methods into a FileSystemCasStorageDriver
- Deprecate and/or remove methods which allow file-level access to the CAS at the level of the CasStorageService and above
