+++ inception/inception-api-dao/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/dao/casstorage/CasStorageServiceImpl.java
@@ -19,40 +19,27 @@ package de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage;
 
 import static de.tudarmstadt.ukp.clarin.webanno.api.CasUpgradeMode.AUTO_CAS_UPGRADE;
 import static de.tudarmstadt.ukp.clarin.webanno.api.CasUpgradeMode.NO_CAS_UPGRADE;
-import static de.tudarmstadt.ukp.clarin.webanno.api.ProjectService.ANNOTATION_FOLDER;
-import static de.tudarmstadt.ukp.clarin.webanno.api.ProjectService.DOCUMENT_FOLDER;
-import static de.tudarmstadt.ukp.clarin.webanno.api.ProjectService.PROJECT_FOLDER;
 import static de.tudarmstadt.ukp.clarin.webanno.api.ProjectService.withProjectLogger;
 import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil.getRealCas;
-import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil.setDocumentId;
 import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil.transferCasOwnershipToCurrentThread;
 import static de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasAccessMode.EXCLUSIVE_WRITE_ACCESS;
 import static de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasAccessMode.SHARED_READ_ONLY_ACCESS;
 import static de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasAccessMode.UNMANAGED_ACCESS;
 import static de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasAccessMode.UNMANAGED_NON_INITIALIZING_ACCESS;
-import static de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.CasMetadataUtils.failOnConcurrentModification;
 import static de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.CasStorageServiceImpl.RepairAndUpgradeFlags.ISOLATED_SESSION;
 import static java.lang.System.currentTimeMillis;
-import static java.nio.file.Files.move;
-import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import static java.util.Collections.newSetFromMap;
 import static java.util.Collections.synchronizedSet;
 import static org.apache.commons.lang3.ArrayUtils.contains;
 
 import java.io.File;
-import java.io.FileFilter;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.WeakHashMap;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.comparator.LastModifiedFileComparator;
 import org.apache.commons.lang3.Validate;
 import org.apache.commons.pool2.impl.GenericKeyedObjectPool;
 import org.apache.commons.pool2.impl.GenericKeyedObjectPoolConfig;
@@ -76,13 +63,10 @@ import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationSchemaService;
 import de.tudarmstadt.ukp.clarin.webanno.api.CasProvider;
 import de.tudarmstadt.ukp.clarin.webanno.api.CasStorageService;
 import de.tudarmstadt.ukp.clarin.webanno.api.CasUpgradeMode;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil;
 import de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasAccessMode;
 import de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasSessionException;
 import de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasStorageServiceAction;
 import de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasStorageServiceLoader;
-import de.tudarmstadt.ukp.clarin.webanno.api.config.RepositoryProperties;
-import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.config.BackupProperties;
 import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.config.CasStorageProperties;
 import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.config.CasStorageServiceAutoConfiguration;
 import de.tudarmstadt.ukp.clarin.webanno.api.event.LayerConfigurationChangedEvent;
@@ -90,7 +74,6 @@ import de.tudarmstadt.ukp.clarin.webanno.diag.CasDoctor;
 import de.tudarmstadt.ukp.clarin.webanno.diag.CasDoctorException;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 import de.tudarmstadt.ukp.clarin.webanno.model.SourceDocument;
-import de.tudarmstadt.ukp.clarin.webanno.support.wicket.WicketUtil;
 
 /**
  * <p>
@@ -105,14 +88,14 @@ public class CasStorageServiceImpl
 
     private final CasDoctor casDoctor;
     private final AnnotationSchemaService schemaService;
-    private final RepositoryProperties repositoryProperties;
-    private final BackupProperties backupProperties;
     private final CasStorageProperties casStorageProperties;
 
     private final GenericKeyedObjectPool<CasKey, CasHolder> exclusiveAccessPool;
     private final Set<CasHolder> exclusiveAccessHolders = synchronizedSet(
             newSetFromMap(new WeakHashMap<>()));
     private final Cache<CasKey, CasHolder> sharedAccessCache;
+
+    private final CasStorageDriver driver;
 
     public static enum RepairAndUpgradeFlags
     {
@@ -129,16 +112,15 @@ public class CasStorageServiceImpl
      * @param aSchemaService
      *            (optional) if present, CAS upgrades can be performed
      */
-    public CasStorageServiceImpl(@Autowired(required = false) CasDoctor aCasDoctor,
+    @Autowired
+    public CasStorageServiceImpl(CasStorageDriver aDriver,
+            @Autowired(required = false) CasDoctor aCasDoctor,
             @Autowired(required = false) AnnotationSchemaService aSchemaService,
-            @Autowired RepositoryProperties aRepositoryProperties,
-            @Autowired CasStorageProperties aCasStorageProperties,
-            @Autowired BackupProperties aBackupProperties)
+            CasStorageProperties aCasStorageProperties)
     {
+        driver = aDriver;
         casDoctor = aCasDoctor;
         schemaService = aSchemaService;
-        repositoryProperties = aRepositoryProperties;
-        backupProperties = aBackupProperties;
         casStorageProperties = aCasStorageProperties;
 
         GenericKeyedObjectPoolConfig<CasHolder> config = new GenericKeyedObjectPoolConfig<>();
@@ -177,15 +159,6 @@ public class CasStorageServiceImpl
 
         if (casDoctor == null) {
             log.info("CAS doctor not available - unable to check/repair CASes");
-        }
-
-        if (backupProperties.getInterval() > 0) {
-            log.info("CAS backups enabled - interval: {}sec  max-backups: {}  max-age: {}sec",
-                    backupProperties.getInterval(), backupProperties.getKeep().getNumber(),
-                    backupProperties.getKeep().getTime());
-        }
-        else {
-            log.info("CAS backups disabled");
         }
 
         log.info("CAS cache size: {} instances", casStorageProperties.getSharedCasCacheSize());
@@ -280,190 +253,6 @@ public class CasStorageServiceImpl
     void writeSerializedCas(CAS aCas, File aFile) throws IOException
     {
         CasPersistenceUtils.writeSerializedCas(aCas, aFile);
-    }
-
-    private void realWriteCas(SourceDocument aDocument, String aUserName, CAS aCas)
-        throws IOException
-    {
-        long t0 = currentTimeMillis();
-
-        analyze(aDocument.getProject(), aDocument.getName(), aDocument.getId(), aUserName, aCas);
-
-        log.debug("Preparing to update annotations for user [{}] on document [{}]({}) " //
-                + "in project [{}]({})", aUserName, aDocument.getName(), aDocument.getId(),
-                aDocument.getProject().getName(), aDocument.getProject().getId());
-
-        File annotationFolder = getAnnotationFolder(aDocument);
-        File currentVersion = new File(annotationFolder, aUserName + ".ser");
-        File oldVersion = new File(annotationFolder, aUserName + ".ser.old");
-
-        // Check if there was a concurrent change to the file on disk
-        if (currentVersion.exists()) {
-            failOnConcurrentModification(aCas, currentVersion, aDocument, aUserName);
-        }
-
-        // Save current version
-        try {
-            // Make a backup of the current version of the file before overwriting
-            if (currentVersion.exists()) {
-                move(currentVersion.toPath(), oldVersion.toPath());
-            }
-
-            // Now write the new version to "<username>.ser" or CURATION_USER.ser
-            long start = currentTimeMillis();
-            setDocumentId(aCas, aUserName);
-            writeSerializedCas(aCas, currentVersion);
-            long duration = currentTimeMillis() - start;
-
-            log.debug(
-                    "Updated annotations for user [{}] on document [{}]({}) in project [{}]({}) in {}ms",
-                    aUserName, aDocument.getName(), aDocument.getId(),
-                    aDocument.getProject().getName(), aDocument.getProject().getId(), duration);
-        }
-        catch (Exception e) {
-            log.error("There was an error while trying to write the CAS to [" + currentVersion
-                    + "] - additional messages follow.");
-            // If this is the first version, there is no old version, so do not restore anything
-            if (!oldVersion.exists()) {
-                log.warn("There is no old version to restore - leaving the current version which "
-                        + "may be corrupt: [{}]", currentVersion);
-                // Now abort anyway
-                throw e;
-            }
-
-            log.error("Restoring previous annotations for user [{}] on document [{}]({}) in " //
-                    + "project [{}]({}) due exception when trying to write new "
-                    + "annotations: [{}]", aUserName, aDocument.getName(), aDocument.getId(),
-                    aDocument.getProject().getName(), aDocument.getProject().getId(), oldVersion);
-            try {
-                move(oldVersion.toPath(), currentVersion.toPath(), REPLACE_EXISTING);
-            }
-            catch (Exception ex) {
-                log.error("Unable to restore previous annotations: [{}]", oldVersion, ex);
-            }
-
-            // Now abort anyway
-            throw e;
-        }
-
-        if (oldVersion.exists() && (currentVersion.length() < oldVersion.length())) {
-            log.debug(
-                    "Annotations truncated for user [{}] on document [{}]({}) in project "
-                            + "[{}]({}): {} -> {} bytes ({} bytes removed)",
-                    aUserName, aDocument.getName(), aDocument.getId(),
-                    aDocument.getProject().getName(), aDocument.getProject().getId(),
-                    oldVersion.length(), currentVersion.length(),
-                    currentVersion.length() - oldVersion.length());
-        }
-
-        // If the saving was successful, we delete the old version
-        if (oldVersion.exists()) {
-            FileUtils.forceDelete(oldVersion);
-        }
-
-        // Update the timestamp in the CAS in case we attempt to save it a second time. This
-        // happens for example in an annotation replacement operation (change layer of existing
-        // annotation) which is implemented as a delete/create operation with an intermediate
-        // save.
-        CasMetadataUtils.addOrUpdateCasMetadata(aCas, currentVersion, aDocument, aUserName);
-
-        manageHistory(currentVersion, aDocument, aUserName);
-
-        WicketUtil.serverTiming("realWriteCas", currentTimeMillis() - t0);
-    }
-
-    private void manageHistory(File aCurrentVersion, SourceDocument aDocument, String aUserName)
-        throws IOException
-    {
-        if (backupProperties.getInterval() <= 0) {
-            return;
-        }
-
-        File annotationFolder = getAnnotationFolder(aDocument);
-
-        // Determine the reference point in time based on the current version
-        long now = aCurrentVersion.lastModified();
-
-        // Get all history files for the current user
-        File[] history = annotationFolder.listFiles(new FileFilter()
-        {
-            private final Matcher matcher = Pattern
-                    .compile(Pattern.quote(aUserName) + "\\.ser\\.[0-9]+\\.bak").matcher("");
-
-            @Override
-            public boolean accept(File aFile)
-            {
-                // Check if the filename matches the pattern given above.
-                return matcher.reset(aFile.getName()).matches();
-            }
-        });
-
-        // Sort the files (oldest one first)
-        Arrays.sort(history, LastModifiedFileComparator.LASTMODIFIED_COMPARATOR);
-
-        // Check if we need to make a new history file
-        boolean historyFileCreated = false;
-        File historyFile = new File(annotationFolder, aUserName + ".ser." + now + ".bak");
-        if (history.length == 0) {
-            // If there is no history yet but we should keep history, then we create a
-            // history file in any case.
-            FileUtils.copyFile(aCurrentVersion, historyFile);
-            historyFileCreated = true;
-        }
-        else {
-            // Check if the newest history file is significantly older than the current one
-            File latestHistory = history[history.length - 1];
-            if (latestHistory.lastModified() + (backupProperties.getInterval() * 1000) < now) {
-                FileUtils.copyFile(aCurrentVersion, historyFile);
-                historyFileCreated = true;
-            }
-        }
-
-        // No history file created, then we can stop here
-        if (!historyFileCreated) {
-            return;
-        }
-
-        // Prune history based on number of backup
-        // The new version is not in the history, so we keep that in any case. That
-        // means we need to keep one less.
-        int toKeep = Math.max(backupProperties.getKeep().getNumber() - 1, 0);
-        if ((backupProperties.getKeep().getNumber() > 0) && (toKeep < history.length)) {
-            // Copy the oldest files to a new array
-            File[] toRemove = new File[history.length - toKeep];
-            System.arraycopy(history, 0, toRemove, 0, toRemove.length);
-
-            // Restrict the history to what is left
-            File[] newHistory = new File[toKeep];
-            if (toKeep > 0) {
-                System.arraycopy(history, toRemove.length, newHistory, 0, newHistory.length);
-            }
-            history = newHistory;
-
-            // Remove these old files
-            for (File file : toRemove) {
-                FileUtils.forceDelete(file);
-
-                log.debug(
-                        "Removed surplus history file [{}] of user [{}] for document {} in project {}",
-                        file.getName(), aUserName, aDocument, aDocument.getProject());
-            }
-        }
-
-        // Prune history based on time
-        if (backupProperties.getKeep().getTime() > 0) {
-            for (File file : history) {
-                if ((file.lastModified() + (backupProperties.getKeep().getTime() * 1000)) < now) {
-                    FileUtils.forceDelete(file);
-
-                    log.debug(
-                            "Removed outdated history file [{}] of user [{}] for "
-                                    + "document [{}]({}) in project [{}]({})",
-                            file.getName(), aUserName, aDocument.getName(), aDocument.getId(),
-                            aDocument.getProject().getName(), aDocument.getProject().getId());
-                }
-            }
-        }
     }
 
     @Override
@@ -624,7 +413,7 @@ public class CasStorageServiceImpl
                 // check for its existence
                 try (WithExclusiveAccess access = new WithExclusiveAccess(aDocument, aUsername)) {
                     casHolder = CasHolder.of(new CasKey(aDocument, aUsername),
-                            () -> readUnmanagedCas(aDocument, aUsername));
+                            () -> driver.readUnmanagedCas(aDocument, aUsername));
                 }
             }
             else {
@@ -740,13 +529,12 @@ public class CasStorageServiceImpl
         String source;
 
         // If the CAS exists on disk already, load it from there
-        File casFile = getCasFile(aDocument, aUsername);
-        if (casFile.exists()) {
+        if (driver.existsCas(aDocument, aUsername)) {
             log.debug("Reading annotation document [{}] ({}) for user [{}] in project [{}] ({})",
                     aDocument.getName(), aDocument.getId(), aUsername,
                     aDocument.getProject().getName(), aDocument.getProject().getId());
 
-            cas = readUnmanagedCas(aDocument, aUsername);
+            cas = driver.readUnmanagedCas(aDocument, aUsername);
             repairAndUpgradeCasIfRequired(aDocument, aUsername, cas, aUpgradeMode,
                     ISOLATED_SESSION);
             source = "disk";
@@ -761,63 +549,20 @@ public class CasStorageServiceImpl
         // If no CAS provider is given, fail
         else {
             throw new FileNotFoundException("CAS file for [" + aDocument.getId() + "," + aUsername
-                    + "] does not exist at [" + casFile + "] and no initializer is specified.");
+                    + "] does not exist and no initializer is specified.");
         }
 
         // Add/update the CAS metadata
-        CasMetadataUtils.addOrUpdateCasMetadata(cas, casFile, aDocument, aUsername);
+        CasMetadataUtils.addOrUpdateCasMetadata(cas,
+                driver.getCasTimestamp(aDocument, aUsername)
+                        .orElseThrow(() -> new IOException(
+                                "Unable to obtain last modified data for annotation document ["
+                                        + aDocument + "] of user [" + aUsername + "]")),
+                aDocument, aUsername);
 
         long duration = currentTimeMillis() - start;
         log.debug("Loaded CAS [{}] [{},{}] from {} in {}ms", cas.hashCode(), aDocument.getId(),
                 aUsername, source, duration);
-
-        return cas;
-    }
-
-    private CAS readUnmanagedCas(SourceDocument aDocument, String aUser) throws IOException
-    {
-        File casFile = getCasFile(aDocument.getProject().getId(), aDocument.getId(), aUser);
-        File oldCasFile = new File(casFile.getPath() + ".old");
-
-        String msgOldExists = "";
-        if (oldCasFile.exists()) {
-            msgOldExists = String.format(
-                    "Existance of temporary annotation file [%s] indicates that a previous "
-                            + "annotation storage process did not successfully complete. Contact "
-                            + "your server administator and request renaming the '.ser.old' file "
-                            + "to '.ser' manually on the command line. Advise the administrator to "
-                            + "check for sufficient disk space and that the application has the "
-                            + "necessary permissions to save files in its data folder.",
-                    oldCasFile);
-        }
-
-        CAS cas;
-        try {
-            cas = WebAnnoCasUtil.createCas();
-        }
-        catch (UIMAException e) {
-            throw new IOException("Unable to create empty CAS", e);
-        }
-
-        if (!casFile.exists()) {
-            throw new FileNotFoundException("Annotation document of user [" + aUser
-                    + "] for source document [" + aDocument.getName() + "] (" + aDocument.getId()
-                    + ") not found in project [" + aDocument.getProject().getName() + "] ("
-                    + aDocument.getProject().getId() + "). " + msgOldExists);
-        }
-
-        try {
-            CasPersistenceUtils.readSerializedCas(cas, casFile);
-            // Add/update the CAS metadata
-            CasMetadataUtils.addOrUpdateCasMetadata(cas, casFile, aDocument, aUser);
-        }
-        catch (Exception e) {
-            throw new IOException("Annotation document of user [" + aUser
-                    + "] for source document [" + aDocument.getName() + "] (" + aDocument.getId()
-                    + ") in project [" + aDocument.getProject().getName() + "] ("
-                    + aDocument.getProject().getId() + ") cannot be read from file [" + casFile
-                    + "]. " + msgOldExists, e);
-        }
 
         return cas;
     }
@@ -828,8 +573,7 @@ public class CasStorageServiceImpl
     {
         try (var logCtx = withProjectLogger(aDocument.getProject());
                 WithExclusiveAccess access = new WithExclusiveAccess(aDocument, aUsername)) {
-            boolean fileWasDeleted = new File(getAnnotationFolder(aDocument), aUsername + ".ser")
-                    .delete();
+            boolean fileWasDeleted = driver.deleteCas(aDocument, aUsername);
 
             // Drop the CAS from the shared CAS it doesn't ghost around. Also set the deleted flag
             // in the holder in case anybody might still be holding on to the holder and needs to
@@ -915,7 +659,7 @@ public class CasStorageServiceImpl
     }
 
     /**
-     * Runs {@link CasDoctor} in anaylsis mode on the given CAS.
+     * Runs {@link CasDoctor} in analysis mode on the given CAS.
      * <p>
      * <b>Note:</b> {@link CasDoctor} is an optional service. If no {@link CasDoctor} implementation
      * is available, this method returns without doing anything.
@@ -962,15 +706,7 @@ public class CasStorageServiceImpl
     @Override
     public File getCasFile(SourceDocument aDocument, String aUser) throws IOException
     {
-        Validate.notNull(aDocument, "Source document must be specified");
-        Validate.notBlank(aUser, "User must be specified");
-
-        return getCasFile(aDocument.getProject().getId(), aDocument.getId(), aUser);
-    }
-
-    public File getCasFile(long aProjectId, long aDocumentId, String aUser) throws IOException
-    {
-        return new File(getAnnotationFolder(aProjectId, aDocumentId), aUser + ".ser");
+        return driver.getCasFile(aDocument, aUser);
     }
 
     @Override
@@ -980,7 +716,7 @@ public class CasStorageServiceImpl
         Validate.notBlank(aUser, "User must be specified");
 
         forceActionOnCas(aDocument, aUser, //
-                (doc, user) -> readUnmanagedCas(doc, user),
+                (doc, user) -> driver.readUnmanagedCas(doc, user),
                 (cas) -> schemaService.upgradeCas(cas, aDocument, aUser), //
                 true);
     }
@@ -1026,7 +762,7 @@ public class CasStorageServiceImpl
         // Ensure that the CAS is not being re-written and temporarily unavailable while we check
         // for its existence
         try (WithExclusiveAccess access = new WithExclusiveAccess(aDocument, aUser)) {
-            return getCasFile(aDocument, aUser).exists();
+            return driver.existsCas(aDocument, aUser);
         }
         catch (IOException e) {
             throw e;
@@ -1173,13 +909,7 @@ public class CasStorageServiceImpl
         // Ensure that the CAS is not being re-written and temporarily unavailable while we check
         // for its existence
         try (WithExclusiveAccess access = new WithExclusiveAccess(aDocument, aUser)) {
-            File casFile = getCasFile(aDocument, aUser);
-            if (!casFile.exists()) {
-                return Optional.empty();
-            }
-            else {
-                return Optional.of(casFile.lastModified());
-            }
+            return driver.getCasTimestamp(aDocument, aUser);
         }
         catch (IOException e) {
             throw e;
@@ -1187,32 +917,6 @@ public class CasStorageServiceImpl
         catch (Exception e) {
             throw new IOException(e);
         }
-    }
-
-    /**
-     * Get the folder where the annotations are stored. Creates the folder if necessary.
-     *
-     * @throws IOException
-     *             if the folder cannot be created.
-     */
-    @Override
-    public File getAnnotationFolder(SourceDocument aDocument) throws IOException
-    {
-        return getAnnotationFolder(aDocument.getProject().getId(), aDocument.getId());
-    }
-
-    /**
-     * Get the folder where the annotations are stored. Creates the folder if necessary.
-     *
-     * @throws IOException
-     *             if the folder cannot be created.
-     */
-    public File getAnnotationFolder(long aProjectId, long aDocumentId) throws IOException
-    {
-        File annotationFolder = new File(repositoryProperties.getPath(), "/" + PROJECT_FOLDER + "/"
-                + aProjectId + "/" + DOCUMENT_FOLDER + "/" + aDocumentId + "/" + ANNOTATION_FOLDER);
-        FileUtils.forceMkdir(annotationFolder);
-        return annotationFolder;
     }
 
     /**
@@ -1256,5 +960,13 @@ public class CasStorageServiceImpl
         // with an updated type system on next access
         sharedAccessCache.asMap().keySet()
                 .removeIf(key -> Objects.equals(key.getProjectId(), aEvent.getProject().getId()));
+    }
+
+    private void realWriteCas(SourceDocument aDocument, String aUserName, CAS aCas)
+        throws IOException
+    {
+        analyze(aDocument.getProject(), aDocument.getName(), aDocument.getId(), aUserName, aCas);
+
+        driver.realWriteCas(aDocument, aUserName, aCas);
     }
 }
