+++ inception/inception-api/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/DocumentService.java
@@ -20,6 +20,7 @@ package de.tudarmstadt.ukp.clarin.webanno.api;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -326,7 +327,7 @@ public interface DocumentService
      */
     boolean existsCas(SourceDocument sourceDocument, String username) throws IOException;
 
-    boolean existsAnnotationCas(AnnotationDocument annotationDocument) throws IOException;
+    boolean existsCas(AnnotationDocument annotationDocument) throws IOException;
 
     /**
      * Export a Serialized CAS annotation document from the file system
@@ -336,7 +337,10 @@ public interface DocumentService
      * @param user
      *            the username.
      * @return the serialized CAS file.
+     * @deprecated Scheduled for removal. To be replaced by a method which takes an
+     *             {@link OutputStream} and then writes the CAS to that stream.
      */
+    @Deprecated
     File getCasFile(SourceDocument document, String user) throws IOException;
 
     /**
